requirements.txt
pandas
streamlit
langchain
langchain-community
chromadb
tiktoken
python-dotenv
sentence-transformers
transformers
langchain-openai